# storage/api-docs/api-docs.json
{"openapi":"3.0.0","info":{"title":"Admin api","version":"1.0.0"},"servers":[{"url":"\/api\/v1"}],"paths":{"\/articles":{"get":{"tags":["Articles"],"operationId":"App\\Http\\Controllers\\Api\\Admin\\ArticleController::index","parameters":[{"name":"with[]","in":"query","description":"With relations","explode":true,"schema":{"type":"array","items":{"type":"string","enum":["mainImage","images"]}}}],"responses":{"200":{"description":"LIST","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]},"post":{"tags":["Articles"],"operationId":"App\\Http\\Controllers\\Api\\Admin\\ArticleController::store","requestBody":{"$ref":"#\/components\/requestBodies\/StoreArticleRequest"},"responses":{"200":{"description":"ITEM","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]}},"\/articles\/{article_id}":{"get":{"tags":["Articles"],"operationId":"App\\Http\\Controllers\\Api\\Admin\\ArticleController::show","parameters":[{"name":"article_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"with[]","in":"query","description":"With relations","explode":true,"schema":{"type":"array","items":{"type":"string","enum":["mainImage","images"]}}}],"responses":{"200":{"description":"ITEM","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]},"put":{"tags":["Articles"],"operationId":"App\\Http\\Controllers\\Api\\Admin\\ArticleController::update","parameters":[{"name":"article_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"$ref":"#\/components\/requestBodies\/UpdateArticleRequest"},"responses":{"200":{"description":"ITEM","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]},"delete":{"tags":["Articles"],"operationId":"App\\Http\\Controllers\\Api\\Admin\\ArticleController::destroy","parameters":[{"name":"article_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"NO CONTENT","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]}},"\/login":{"post":{"tags":["General"],"summary":"login","description":"Get a JWT via credentials","operationId":"App\\Http\\Controllers\\Api\\Admin\\Auth\\LoginController::login","requestBody":{"request":"LoginRequest","description":"Login Request","required":true,"content":{"application\/json":{"schema":{"required":["email","password"],"properties":{"email":{"description":"login\n\nGet a JWT token via given credentials.","type":"string","format":"email","maximum":255},"password":{"description":"login\n\nGet a JWT token via given credentials.","type":"string","minimum":6}},"type":"object"},"example":{"email":"test@test.com","password":"123456"}}}},"responses":{"200":{"description":"Success login","content":{"application\/json":{"schema":{}}}},"401":{"description":"Invalid credentials","content":{"application\/json":{"schema":{}}}},"422":{"description":"Validation errors","content":{"application\/json":{"schema":{}}}}}}},"\/logout":{"post":{"tags":["General"],"summary":"logout","description":"Logout","operationId":"App\\Http\\Controllers\\Api\\Admin\\Auth\\LoginController::logout","responses":{"204":{"description":"Success login","content":{"application\/json":{"schema":{}}}},"401":{"description":"Unauthorized","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]}},"\/refresh":{"post":{"tags":["General"],"summary":"refresh","description":"Refresh JWT token","operationId":"App\\Http\\Controllers\\Api\\Admin\\Auth\\LoginController::refresh","responses":{"200":{"description":"Success login","content":{"application\/json":{"schema":{}}}},"401":{"description":"Unauthorized","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]}},"\/categories":{"get":{"tags":["Categories"],"operationId":"App\\Http\\Controllers\\Api\\Admin\\CategoryController::index","responses":{"200":{"description":"LIST","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]},"post":{"tags":["Categories"],"operationId":"App\\Http\\Controllers\\Api\\Admin\\CategoryController::store","requestBody":{"$ref":"#\/components\/requestBodies\/StoreCategoryRequest"},"responses":{"200":{"description":"ITEM","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]}},"\/categories\/{category_id}":{"get":{"tags":["Categories"],"operationId":"App\\Http\\Controllers\\Api\\Admin\\CategoryController::show","parameters":[{"name":"category_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"ITEM","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]},"put":{"tags":["Categories"],"operationId":"App\\Http\\Controllers\\Api\\Admin\\CategoryController::update","parameters":[{"name":"category_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"$ref":"#\/components\/requestBodies\/UpdateCategoryRequest"},"responses":{"200":{"description":"ITEM","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]},"delete":{"tags":["Categories"],"operationId":"App\\Http\\Controllers\\Api\\Admin\\CategoryController::destroy","parameters":[{"name":"category_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"NO CONTENT","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]}},"\/images":{"get":{"tags":["Images"],"operationId":"App\\Http\\Controllers\\Api\\Admin\\ImageController::index","parameters":[{"name":"filter[type]","in":"query","description":"Filter by image type","schema":{"type":"string","enum":["unknown","icon"]}}],"responses":{"200":{"description":"LIST","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]},"post":{"tags":["Images"],"operationId":"App\\Http\\Controllers\\Api\\Admin\\ImageController::store","requestBody":{"$ref":"#\/components\/requestBodies\/StoreImagesRequest"},"responses":{"200":{"description":"LIST","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]}},"\/images\/{image_id}":{"delete":{"tags":["Images"],"operationId":"App\\Http\\Controllers\\Api\\Admin\\ImageController::destroy","parameters":[{"name":"image_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"NO CONTENT","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]}},"\/posts":{"get":{"tags":["Posts"],"operationId":"App\\Http\\Controllers\\Api\\Admin\\PostController::index","parameters":[{"name":"with[]","in":"query","description":"With relations","explode":true,"schema":{"type":"array","items":{"type":"string","enum":["mainImage","images"]}}}],"responses":{"200":{"description":"LIST","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]},"post":{"tags":["Posts"],"operationId":"App\\Http\\Controllers\\Api\\Admin\\PostController::store","requestBody":{"$ref":"#\/components\/requestBodies\/StorePostRequest"},"responses":{"200":{"description":"ITEM","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]}},"\/posts\/{post_id}":{"get":{"tags":["Posts"],"operationId":"App\\Http\\Controllers\\Api\\Admin\\PostController::show","parameters":[{"name":"post_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"with[]","in":"query","description":"With relations","explode":true,"schema":{"type":"array","items":{"type":"string","enum":["mainImage","images"]}}}],"responses":{"200":{"description":"ITEM","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]},"put":{"tags":["Posts"],"operationId":"App\\Http\\Controllers\\Api\\Admin\\PostController::update","parameters":[{"name":"post_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"$ref":"#\/components\/requestBodies\/UpdatePostRequest"},"responses":{"200":{"description":"ITEM","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]},"delete":{"tags":["Posts"],"operationId":"App\\Http\\Controllers\\Api\\Admin\\PostController::destroy","parameters":[{"name":"post_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"NO CONTENT","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]}},"\/reviews":{"get":{"tags":["Reviews"],"operationId":"App\\Http\\Controllers\\Api\\Admin\\ReviewController::index","parameters":[{"name":"with[]","in":"query","description":"With relations","explode":true,"schema":{"type":"array","items":{"type":"string","enum":["avatar","images"]}}}],"responses":{"200":{"description":"LIST","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]},"post":{"tags":["Reviews"],"operationId":"App\\Http\\Controllers\\Api\\Admin\\ReviewController::store","requestBody":{"$ref":"#\/components\/requestBodies\/StoreReviewRequest"},"responses":{"200":{"description":"ITEM","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]}},"\/reviews\/{review_id}":{"get":{"tags":["Reviews"],"operationId":"App\\Http\\Controllers\\Api\\Admin\\ReviewController::show","parameters":[{"name":"review_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"with[]","in":"query","description":"With relations","explode":true,"schema":{"type":"array","items":{"type":"string","enum":["avatar","images"]}}}],"responses":{"200":{"description":"ITEM","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]},"put":{"tags":["Reviews"],"operationId":"App\\Http\\Controllers\\Api\\Admin\\ReviewController::update","parameters":[{"name":"review_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"$ref":"#\/components\/requestBodies\/UpdateReviewRequest"},"responses":{"200":{"description":"ITEM","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]},"delete":{"tags":["Reviews"],"operationId":"App\\Http\\Controllers\\Api\\Admin\\ReviewController::destroy","parameters":[{"name":"review_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"NO CONTENT","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]}},"\/tours":{"get":{"tags":["Tours"],"operationId":"App\\Http\\Controllers\\Api\\Admin\\TourController::index","parameters":[{"name":"with[]","in":"query","description":"With relations","explode":true,"schema":{"type":"array","items":{"type":"string","enum":["mainImage","images"]}}}],"responses":{"200":{"description":"LIST","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]},"post":{"tags":["Tours"],"operationId":"App\\Http\\Controllers\\Api\\Admin\\TourController::store","requestBody":{"$ref":"#\/components\/requestBodies\/StoreTourRequest"},"responses":{"200":{"description":"ITEM","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]}},"\/tours\/{tour_id}":{"get":{"tags":["Tours"],"operationId":"App\\Http\\Controllers\\Api\\Admin\\TourController::show","parameters":[{"name":"tour_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"with[]","in":"query","description":"With relations","explode":true,"schema":{"type":"array","items":{"type":"string","enum":["mainImage","images"]}}}],"responses":{"200":{"description":"ITEM","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]},"put":{"tags":["Tours"],"operationId":"App\\Http\\Controllers\\Api\\Admin\\TourController::update","parameters":[{"name":"tour_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"$ref":"#\/components\/requestBodies\/UpdateTourRequest"},"responses":{"200":{"description":"ITEM","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]},"delete":{"tags":["Tours"],"operationId":"App\\Http\\Controllers\\Api\\Admin\\TourController::destroy","parameters":[{"name":"tour_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"NO CONTENT","content":{"application\/json":{"schema":{}}}}},"security":[{"bearer":[]}]}}},"components":{"parameters":{"paginate":{"name":"paginate","in":"query","description":"Enable pagination","required":false,"schema":{"type":"boolean","default":true}}},"requestBodies":{"StoreArticleRequest":{"description":"Store Article Request","required":true,"content":{"application\/json":{"schema":{"required":["title","short_description","description","main_image_id"],"properties":{"meta_title":{"type":"string","maxLength":255},"meta_keywords":{"type":"string","maxLength":1000},"meta_description":{"type":"string","maxLength":1000},"main_image_id":{"type":"integer"},"title":{"type":"string","maxLength":255},"short_description":{"type":"string"},"description":{"type":"string"},"type":{"type":"string","enum":["country","direction"]},"relations":{"properties":{"images":{"type":"array","items":{"type":"integer"}}},"type":"object"}},"type":"object"}}}},"UpdateArticleRequest":{"description":"Update Article Request","required":true,"content":{"application\/json":{"schema":{"properties":{"meta_title":{"type":"string","maxLength":255},"meta_keywords":{"type":"string","maxLength":1000},"meta_description":{"type":"string","maxLength":1000},"main_image_id":{"type":"integer"},"title":{"type":"string","maxLength":255},"short_description":{"type":"string"},"description":{"type":"string"},"type":{"type":"string","enum":["country","direction"]},"relations":{"properties":{"images":{"type":"array","items":{"type":"integer"}}},"type":"object"}},"type":"object"}}}},"StoreCategoryRequest":{"description":"Store Category Request","required":true,"content":{"application\/json":{"schema":{"required":["name"],"properties":{"name":{"type":"string","maxLength":255}},"type":"object"}}}},"UpdateCategoryRequest":{"description":"Update Category Request","required":true,"content":{"application\/json":{"schema":{"required":["name"],"properties":{"name":{"type":"string","maxLength":255}},"type":"object"}}}},"StoreImagesRequest":{"description":"Store Images Request","required":true,"content":{"multipart\/form-data":{"schema":{"properties":{"images":{"description":"\u0412 \u0434\u043e\u043a\u0443\u043c\u0435\u043d\u0442\u0430\u0446\u0438\u0438 \u043d\u0435 \u0440\u0430\u0431\u043e\u0442\u0430\u0435\u0442 \u043c\u0430\u0441\u0441\u043e\u0432\u0430\u044f \u0437\u0430\u0433\u0440\u0443\u0437\u043a\u0430 \u0444\u0430\u0439\u043b\u043e\u0432","type":"array","items":{"type":"string","format":"binary"}},"width":{"type":"integer","maximum":1920,"minimum":100},"height":{"type":"integer","maximum":1920,"minimum":100},"type":{"type":"string","default":"unknown","enum":["unknown","icon"]}},"type":"object"}}}},"StorePostRequest":{"description":"Store Post Request","required":true,"content":{"application\/json":{"schema":{"required":["title","short_description","description","main_image_id"],"properties":{"category_id":{"type":"integer"},"meta_title":{"type":"string","maxLength":255},"meta_keywords":{"type":"string","maxLength":1000},"meta_description":{"type":"string","maxLength":1000},"main_image_id":{"type":"integer"},"title":{"type":"string","maxLength":255},"short_description":{"type":"string","maxLength":255},"description":{"type":"string"},"relations":{"properties":{"images":{"type":"array","items":{"type":"integer"}}},"type":"object"}},"type":"object"}}}},"UpdatePostRequest":{"description":"Update Post Request","required":true,"content":{"application\/json":{"schema":{"properties":{"category_id":{"type":"integer"},"meta_title":{"type":"string","maxLength":255},"meta_keywords":{"type":"string","maxLength":1000},"meta_description":{"type":"string","maxLength":1000},"main_image_id":{"type":"integer"},"title":{"type":"string","maxLength":255},"short_description":{"type":"string","maxLength":255},"description":{"type":"string"},"relations":{"properties":{"images":{"type":"array","items":{"type":"integer"}}},"type":"object"}},"type":"object"}}}},"StoreReviewRequest":{"description":"Store Review Request","required":true,"content":{"application\/json":{"schema":{"required":["avatar_id","name","date","body","rating"],"properties":{"avatar_id":{"type":"integer"},"name":{"type":"string","maxLength":255},"date":{"type":"string","format":"date"},"body":{"type":"string"},"rating":{"type":"integer","maximum":5,"minimum":0},"link":{"type":"string"},"icon_id":{"type":"integer"},"relations":{"properties":{"images":{"type":"array","items":{"type":"integer"}}},"type":"object"}},"type":"object"}}}},"UpdateReviewRequest":{"description":"Update Review Request","required":true,"content":{"application\/json":{"schema":{"properties":{"avatar_id":{"type":"integer"},"name":{"type":"string","maxLength":255},"date":{"type":"string","format":"date"},"body":{"type":"string"},"rating":{"type":"integer","maximum":5,"minimum":0},"link":{"type":"string"},"icon_id":{"type":"integer"},"relations":{"properties":{"images":{"type":"array","items":{"type":"integer"}}},"type":"object"}},"type":"object"}}}},"StoreTourRequest":{"description":"Store Tour Request","required":true,"content":{"application\/json":{"schema":{"required":["country","city","duration","tourists_number","departure_date","food","note","main_image_id","description","price","is_hot"],"properties":{"meta_title":{"type":"string","maxLength":255},"meta_keywords":{"type":"string","maxLength":1000},"meta_description":{"type":"string","maxLength":1000},"country":{"type":"string","maxLength":255},"city":{"type":"string","maxLength":255},"hotel":{"type":"string","maxLength":255},"duration":{"type":"integer","maximum":255},"tourists_number":{"type":"integer","maximum":255},"departure_date":{"type":"string","format":"date"},"food":{"type":"string","maxLength":255},"note":{"type":"string","maxLength":255},"main_image_id":{"type":"integer"},"description":{"type":"string"},"price":{"type":"number","format":"float"},"is_hot":{"type":"boolean"},"relations":{"properties":{"images":{"type":"array","items":{"type":"integer"}}},"type":"object"}},"type":"object"}}}},"UpdateTourRequest":{"description":"Update Tour Request","required":true,"content":{"application\/json":{"schema":{"properties":{"meta_title":{"type":"string","maxLength":255},"meta_keywords":{"type":"string","maxLength":1000},"meta_description":{"type":"string","maxLength":1000},"country":{"type":"string","maxLength":255},"city":{"type":"string","maxLength":255},"hotel":{"type":"string","maxLength":255},"duration":{"type":"integer","maximum":255},"tourists_number":{"type":"integer","maximum":255},"departure_date":{"type":"string","format":"date"},"food":{"type":"string","maxLength":255},"note":{"type":"string","maxLength":255},"main_image_id":{"type":"integer"},"description":{"type":"string"},"price":{"type":"number","format":"float"},"is_hot":{"type":"boolean"},"relations":{"properties":{"images":{"type":"array","items":{"type":"integer"}}},"type":"object"}},"type":"object"}}}}},"securitySchemes":{"bearer":{"type":"http","description":"JWT security scheme","name":"Authorization","in":"header","scheme":"bearer","bearerFormat":"JWT"}}}}